# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "liber API",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "2"
  },
  "paths": {
    "/api/v2/bookmarks": {
      "get": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "list_bookmarks",
        "parameters": [
          {
            "name": "q",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "sort",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "folder",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "deep",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged bookmark list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad sort mode",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::bookmarks"
        ],
        "operationId": "add_bookmark",
        "requestBody": {
          "description": "Bookmark to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Duplicate accepted idempotently",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "201": {
            "description": "Bookmark created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiBookmark"
                }
              }
            }
          },
          "400": {
            "description": "Missing url",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate url",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/rules": {
      "get": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "list_rules",
        "responses": {
          "200": {
            "description": "Rules with applied counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "add_rule_ep",
        "requestBody": {
          "description": "Rule as {pattern, tags?, folder?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Rule created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad pattern",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Rule exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/rules/apply": {
      "post": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "apply_rules_ep",
        "requestBody": {
          "description": "Scope as {id?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Applied count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/rules/learn": {
      "get": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "learn_rules",
        "parameters": [
          {
            "name": "min",
            "in": "query",
            "description": "Minimum host count, at least 2",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Host suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "learn_create",
        "requestBody": {
          "description": "Threshold as {min?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created plus applied counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/rules/{id}": {
      "put": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "edit_rule_ep",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Rule id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Partial rule plus {reapply?}",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Rule with reapplied count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No such rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "crate::rules"
        ],
        "operationId": "delete_rule_ep",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Rule id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted rule id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "No such rule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "archive": {
            "type": "boolean"
          },
          "confirm_dup": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "folder": {
            "type": [
              "string",
              "null"
            ]
          },
          "markdown": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ApiAttachment": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ApiBookmark": {
        "type": "object",
        "required": [
          "uuid",
          "url",
          "title",
          "created_at",
          "updated_at",
          "has_markdown",
          "has_archive"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiAttachment"
            }
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "folder": {
            "type": "string"
          },
          "has_archive": {
            "type": "boolean"
          },
          "has_markdown": {
            "type": "boolean"
          },
          "last_opened_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "open_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "short_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "required": [
          "total",
          "page",
          "per_page",
          "bookmarks"
        ],
        "properties": {
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiBookmark"
            }
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}